requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
